Move useDebounceEffect
