+++ src/js/base/util.tsx
@@ -1,3 +1,5 @@
+import { useEffect, DependencyList } from "react"
+
 export function safeParseFloat(input: string | null | undefined) {
   if (input) {
     return parseFloat(input)
@@ -10,4 +12,18 @@ export function safeParseInt(input: string | null | undefined) {
     return parseInt(input, 10)
   }
   return null
+}
+
+export function useDebounceEffect(fn: () => void, waitTime: number, deps?: DependencyList) {
+  useEffect(() => {
+    const t = setTimeout(() => {
+      // eslint-disable-next-line prefer-spread
+      fn.apply(undefined, deps)
+    }, waitTime)
+
+    return () => {
+      clearTimeout(t)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [deps])
 }
