--- src/js/users/use_debounce_effect.ts
@@ -1,15 +1,0 @@
-import { useEffect, DependencyList } from "react"
-
-export default function useDebounceEffect(fn: () => void, waitTime: number, deps?: DependencyList) {
-  useEffect(() => {
-    const t = setTimeout(() => {
-      // eslint-disable-next-line prefer-spread
-      fn.apply(undefined, deps)
-    }, waitTime)
-
-    return () => {
-      clearTimeout(t)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [deps])
-}
